Add continuous statistics to a Likert table

This function converts Likert-scales into a numeric score and computes continuous statistics based on this score.

add_continuous_stat(x, ...)

# S3 method for tbl_likert
add_continuous_stat(
  x,
  statistic = "{mean}",
  digits = NULL,
  col_label = NULL,
  footnote = FALSE,
  last = TRUE,
  score_values = NULL,
  stat_col_name = NULL,
  ...
)

Arguments

x

Object with class tbl_likert from the tbl_likert() function

...

not used

statistic

String or formula indicating the statistic to be reported. Default is the mean score. Other possible continuous statistics are described in gtsummary::tbl_summary() help page, section statistic argument.

digits

Formula or list of formulas indicating how to display the computed statistics, see gtsummary::tbl_summary() help page

col_label

String indicating the column label. Default is generated from statistic.

footnote

Logical argument indicating whether to print a footnote clarifying the statistics presented. Default is FALSE

last

Logical indicator to include the new column last in table. Default is TRUE

score_values

Vector indicating the numeric value of each factor level. Default is 1:n where n indicates the number of levels.

stat_col_name

Optional string indicating the name of the new column added to x$table_body

See also