Add column with N to a Likert table

Add column with N to a Likert table

# S3 method for tbl_likert
add_n(
  x,
  statistic = "{n}",
  col_label = "**N**",
  footnote = FALSE,
  last = FALSE,
  ...
)

Arguments

x

Object with class tbl_likert from the tbl_likert() function

statistic

String indicating the statistic to report. Default is the number of non-missing observation for each variable, statistic = "{n}". Other statistics available to report include:

"{N}" total number of observations,

"{n}" number of non-missing observations,

"{n_miss}" number of missing observations,

"{p}" percent non-missing data,

"{p_miss}" percent missing data The argument uses glue::glue syntax and multiple statistics may be reported, e.g. statistic = "{n} / {N} ({p}%)"

col_label

String indicating the column label. Default is "**N**"

footnote

Logical argument indicating whether to print a footnote clarifying the statistics presented. Default is FALSE

last

Logical indicator to include N column last in table. Default is FALSE, which will display N column first.

...

Not used

See also