Start a New Biostatistics project

Creates a directory with the essential files for a new project. The function can be used on existing project directories as well. This is a thin wrapper for starter::create_project() that sets the default template to template = hotfun::project_template

create_bst_project(
  path,
  path_data = NULL,
  template = NULL,
  git = NA,
  renv = TRUE,
  overwrite = NA,
  open = interactive()
)

Arguments

path

A path. If it exists, it is used. If it does not exist, it is created.

path_data

A path. The directory where the secure data exist. Default is NULL. When supplied, a symbolic link to data folder will be created.

template

Specifies template for starter::create_project(template=). Default is the template in bstfun::project_templates whose name matches the lowercase system username (Sys.info()[["user"]]), if it exists; otherwise, bstfun::project_templates[["default"]]

git

Logical indicating whether to create Git repository. Default is TRUE When NA, user will be prompted whether to initialise Git repo.

renv

Logical indicating whether to add renv to a project. Default is TRUE. When NA user is asked interactively for preference.

overwrite

Logical indicating whether to overwrite existing files if they exist. Options are TRUE, FALSE, and NA (aka ask interactively). Default is NA

open

Logical indicating whether to open new project in fresh RStudio session

Examples