Moves p-value footnotes to individual p-values

This function converts a gtsummary object to the output type indicated in printer=: your object will no longer be of class gtsummary and you can no longer use gtsummary functions to modify the object.

gts_add_p_footnotes(x, printer = c("gt", "flextable"), index_start = 2)

Arguments

x
object with class "tbl_summary" created from the gtsummary package

printer
String indicating the output format. Must be one of "gt", or "flextable"

index_start
for flextable output, this is the number the footnotes begin counting from

Examples

library(gtsummary)
trial %>%
  select(trt, age, grade) %>%
  tbl_summary(by = trt) %>%
  add_p() %>%
  gts_add_p_footnotes()

#> Warning: The `.dots` argument of `group_by()` is deprecated as of dplyr 1.0.0.

#> <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #wlefoodfie .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #wlefoodfie .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #wlefoodfie .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #wlefoodfie .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 4px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #wlefoodfie .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #wlefoodfie .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #wlefoodfie .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #wlefoodfie .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #wlefoodfie .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #wlefoodfie .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #wlefoodfie .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #wlefoodfie .gt_group_heading {
#>   padding: 8px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #wlefoodfie .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #wlefoodfie .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #wlefoodfie .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #wlefoodfie .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #wlefoodfie .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 12px;
#> }
#> 
#> #wlefoodfie .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #wlefoodfie .gt_first_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #wlefoodfie .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #wlefoodfie .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #wlefoodfie .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #wlefoodfie .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #wlefoodfie .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #wlefoodfie .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #wlefoodfie .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #wlefoodfie .gt_sourcenote {
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #wlefoodfie .gt_left {
#>   text-align: left;
#> }
#> 
#> #wlefoodfie .gt_center {
#>   text-align: center;
#> }
#> 
#> #wlefoodfie .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #wlefoodfie .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #wlefoodfie .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #wlefoodfie .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #wlefoodfie .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #wlefoodfie .gt_footnote_marks {
#>   font-style: italic;
#>   font-size: 65%;
#> }
#> </style>
#> <div id="wlefoodfie" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;"><table class="gt_table">
#>   
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1"><strong>Characteristic</strong></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1"><strong>Drug A</strong>, N = 98<sup class="gt_footnote_marks">1</sup></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1"><strong>Drug B</strong>, N = 102<sup class="gt_footnote_marks">1</sup></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1"><strong>p-value</strong></th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr>
#>       <td class="gt_row gt_left">Age</td>
#>       <td class="gt_row gt_center">46 (37, 59)</td>
#>       <td class="gt_row gt_center">48 (39, 56)</td>
#>       <td class="gt_row gt_center">0.7<sup class="gt_footnote_marks">2</sup></td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="text-align: left; text-indent: 10px;">Unknown</td>
#>       <td class="gt_row gt_center">7</td>
#>       <td class="gt_row gt_center">4</td>
#>       <td class="gt_row gt_center"></td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left">Grade</td>
#>       <td class="gt_row gt_center"></td>
#>       <td class="gt_row gt_center"></td>
#>       <td class="gt_row gt_center">0.9<sup class="gt_footnote_marks">3</sup></td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="text-align: left; text-indent: 10px;">I</td>
#>       <td class="gt_row gt_center">35 (36%)</td>
#>       <td class="gt_row gt_center">33 (32%)</td>
#>       <td class="gt_row gt_center"></td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="text-align: left; text-indent: 10px;">II</td>
#>       <td class="gt_row gt_center">32 (33%)</td>
#>       <td class="gt_row gt_center">36 (35%)</td>
#>       <td class="gt_row gt_center"></td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="text-align: left; text-indent: 10px;">III</td>
#>       <td class="gt_row gt_center">31 (32%)</td>
#>       <td class="gt_row gt_center">33 (32%)</td>
#>       <td class="gt_row gt_center"></td>
#>     </tr>
#>   </tbody>
#>   
#>   <tfoot>
#>     <tr class="gt_footnotes">
#>       <td colspan="4">
#>         <p class="gt_footnote">
#>           <sup class="gt_footnote_marks">
#>             <em>1</em>
#>           </sup>
#>            
#>           Median (IQR); n (%)
#>           <br />
#>         </p>
#>         <p class="gt_footnote">
#>           <sup class="gt_footnote_marks">
#>             <em>2</em>
#>           </sup>
#>            
#>           Wilcoxon rank sum test
#>           <br />
#>         </p>
#>         <p class="gt_footnote">
#>           <sup class="gt_footnote_marks">
#>             <em>3</em>
#>           </sup>
#>            
#>           Pearson's Chi-squared test
#>           <br />
#>         </p>
#>       </td>
#>     </tr>
#>   </tfoot>
#> </table></div>