Likert Summary

tbl_likert(
  data,
  label = NULL,
  statistic = NULL,
  digits = NULL,
  include = everything(),
  sort = c("default", "ascending", "descending")
)

# S3 method for tbl_likert
add_n(x, ...)

Arguments

data
A data frame

label
List of formulas specifying variables labels, e.g. list(age ~ "Age", stage ~ "Path T Stage"). If a variable's label is not specified here, the label attribute (attr(data$age, "label")) is used. If attribute label is NULL, the variable name will be used.

statistic
List of formulas specifying types of summary statistics to display for each variable. The default is list(all_continuous() ~ "{median} ({p25}, {p75})", all_categorical() ~ "{n} ({p}%)"). See below for details.

digits
List of formulas specifying the number of decimal places to round continuous summary statistics. If not specified, tbl_summary guesses an appropriate number of decimals to round statistics. When multiple statistics are displayed for a single variable, supply a vector rather than an integer. For example, if the statistic being calculated is "{mean} ({sd})" and you want the mean rounded to 1 decimal place, and the SD to 2 use digits = list(age ~ c(1, 2)). User may also pass a styling function: digits = age ~ style_sigfig

include
variables to include in the summary table. Default is everything()

sort
Sort table based on mean scores? Must be one of c("default", "ascending", "descending")

x
Table of class 'tbl_likert'

...
not used

Examples

df <-
  tibble::tibble(
    f1 =
      sample.int(100, n = 3, replace = TRUE) %>%
      factor(levels = 1:3, labels = c("bad", "meh", "good")),
    f2 =
      sample.int(100, n = 3, replace = TRUE) %>%
      factor(levels = 1:3, labels = c("bad", "meh", "good")),
  )

tbl_likert(df) %>%
  add_n() %>%
  gtsummary::as_kable()

#> <table>
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;"> **Characteristic** </th>
#>    <th style="text-align:left;"> **N** </th>
#>    <th style="text-align:left;"> **bad** </th>
#>    <th style="text-align:left;"> **meh** </th>
#>    <th style="text-align:left;"> **good** </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> f1 </td>
#>    <td style="text-align:left;"> 100 </td>
#>    <td style="text-align:left;"> 33 (33%) </td>
#>    <td style="text-align:left;"> 35 (35%) </td>
#>    <td style="text-align:left;"> 32 (32%) </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> f2 </td>
#>    <td style="text-align:left;"> 100 </td>
#>    <td style="text-align:left;"> 36 (36%) </td>
#>    <td style="text-align:left;"> 27 (27%) </td>
#>    <td style="text-align:left;"> 37 (37%) </td>
#>   </tr>
#> </tbody>
#> </table>