Set custom gtsummary themes

This is a place for any member of the MSK community to add a personal gtsummary theme. Reach out if you're interested in adding yours!

theme_gtsummary_msk(
  name = c("hot", "karissa", "ally", "mauguen", "esther", "curry", "lavery", "meier",
    "leej", "whitingk", "mauguena", "drille", "currym1", "laveryj", "hsum1", "leej22")
)

Arguments

name

string indicating the custom theme to set.

Details

Visit the gtsummary themes vignette for a full list of preferences that can be set.

See also

Examples

theme_gtsummary_msk("hot")
#> Setting theme `H.O.T.`