Results from a simulated study of two chemotherapy agents: Drug A and Drug B

A dataset containing the baseline characteristics of 200 patients who received Drug A or Drug B. Dataset also contains the outcome of tumor response to the treatment.

trial

Format

A data frame with 200 rows--one row per patient

trt

Chemotherapy Treatment

age

Age, yrs

marker

Marker Level, ng/mL

stage

T Stage

grade

Grade

response

Tumor Response

death

Patient Died

ttdeath

Months to Death/Censor