Write a template file

Rather than using create_bst_project() to start a new project folder, you may use use_bst_file() to write a single file from any project template. The functions use_bst_gitignore() and use_bst_readme() are shortcuts for use_bst_file("gitignore") and use_bst_file("readme").

use_bst_file(
  name = NULL,
  filename = NULL,
  template = NULL,
  open = interactive()
)

use_bst_gitignore(filename = NULL, template = NULL)

use_bst_readme(filename = NULL, template = NULL)

Arguments

name
Name of file to write. Not sure of the files available to you? Run the function without specifying a name, and all files available within the template will print.

filename
Optional argument to specify the name of the file to be written. Paths/filename is relative to project base (e.g. here::here())

template
A project template. See vignette for details.

open
If TRUE, opens the new file.

See also

Examples

if (FALSE) {
# create gitignore file
use_project_file("gitignore")
use_project_gitignore()

# create README.md file
use_project_file("readme")
use_project_readme()
}